+++ packages/common/package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "@pixuli/core": "workspace:*",
+    "@pixuli/ui": "workspace:*",
     "lucide-react": "^0.556.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
