+++ packages/common/src/components/features/version-info/web/VersionInfoModal.tsx
@@ -1,4 +1,4 @@
-import { useEscapeKey } from '../../../../hooks';
+import { useEscapeKey } from '@pixuli/ui';
 import {
   Calendar,
   Code,
