+++ packages/common/src/components/layout/empty-state/web/index.ts
@@ -1,2 +1,2 @@
 export { default as EmptyState } from './EmptyState';
-export type { EmptyStateProps } from '../common/types';
+export type { EmptyStateProps } from '@pixuli/ui/native';
