+++ packages/common/src/components/ui/keyboard-help/web/KeyboardHelpModal.tsx
@@ -1,7 +1,7 @@
 // TODO: 自定义快捷键
 import { Command, Keyboard, RefreshCw, X, Zap } from 'lucide-react';
 import React from 'react';
-import { useEscapeKey } from '../../../../hooks';
+import { useEscapeKey } from '@pixuli/ui';
 import { defaultTranslate } from '../../../../locales';
 import './KeyboardHelpModal.css';
