+++ packages/common/src/index.native.ts
@@ -13,13 +13,8 @@ export * from '@pixuli/core/types';
 // ==================== Services 导出 ====================
 export * from './services';
 
-// ==================== React Native 专用组件导出 ====================
-// EmptyState 组件（React Native 版本）
-export * from './components/layout/empty-state/native';
-// Demo 组件（React Native 版本）
-export * from './components/dev/demo/native';
-// VersionInfoModal 组件（React Native 版本）
-export * from './components/features/version-info/native';
+// ==================== React Native 专用组件（@pixuli/ui/native） ====================
+export * from '@pixuli/ui/native';
 
 // ==================== 语言包导出 ====================
 export * from './locales';
