+++ packages/common/src/index.ts
@@ -13,8 +13,8 @@ export * from './services';
 // ==================== 工具函数导出 ====================
 export * from './utils';
 
-// ==================== Hooks 导出 ====================
-export * from './hooks';
+// ==================== Hooks 导出（@pixuli/ui） ====================
+export * from '@pixuli/ui';
 
 // ==================== 组件导出 ====================
 // 图片相关组件
