+++ packages/ui/src/dev/demo/locales/index.ts
