+++ packages/ui/src/features/version-info/locales/en-US.json
