+++ packages/ui/src/features/version-info/locales/index.ts
