+++ packages/ui/src/features/version-info/locales/zh-CN.json
