+++ packages/ui/src/layout/empty-state/common/types.ts
