+++ packages/ui/src/layout/empty-state/locales/en-US.json
