+++ packages/ui/src/layout/empty-state/locales/index.ts
