+++ packages/ui/src/layout/empty-state/locales/zh-CN.json
