+++ packages/ui/src/web/index.ts
@@ -1,0 +1,5 @@
+/**
+ * @pixuli/ui — Web / Desktop 入口
+ * L1/L2 组件将在 REF-203 迁入
+ */
+export * from '../hooks';
